requirements.txt
Flask>=2.3,<3.1
pandas
openpyxl
validators
sendgrid
python-dotenv
imaplib2

